+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Diagnostic Settings entities and relationships for Azure API Management
+  Services
+
 ## 5.10.1 - 2020-01-15
 
 ### Fixed